review fixes
* HashHelpers to StringDictionarySlim
